requirements.txt
pdfplumber
pdf2image
pytesseract
google-generativeai
pandas
openpyxl
tkinter
Pillow
pyttsx3